@@ -17,7 +17,7 @@
         "allow-plugins": false
     },
     "require": {
-        "php": "^8.3",
+        "php": "^8.4",
         "doctrine/lexer": "^3.0",
         "meyfa/php-svg": "^0.16.0",
         "symfony/console": "^7.1"
